Move Composer event listeners into namespace
