+++ src/ComposerIntegration/PieComposerFactory.php
@@ -11,6 +11,8 @@ use Composer\IO\IOInterface;
 use Composer\PartialComposer;
 use Composer\Util\Filesystem;
 use Composer\Util\ProcessExecutor;
+use Php\Pie\ComposerIntegration\Listeners\OverrideWindowsUrlInstallListener;
+use Php\Pie\ComposerIntegration\Listeners\RemoveUnrelatedInstallOperations;
 use Php\Pie\ExtensionType;
 use Php\Pie\Platform;
 use Psr\Container\ContainerInterface;
