+++ test/unit/ComposerIntegration/Listeners/RemoveUnrelatedInstallOperationsTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Php\PieUnitTest\ComposerIntegration;
+namespace Php\PieUnitTest\ComposerIntegration\Listeners;
 
 use Composer\Composer;
 use Composer\DependencyResolver\Operation\InstallOperation;
@@ -15,7 +15,7 @@ use Composer\IO\IOInterface;
 use Composer\Package\CompletePackage;
 use Php\Pie\ComposerIntegration\PieComposerRequest;
 use Php\Pie\ComposerIntegration\PieOperation;
-use Php\Pie\ComposerIntegration\RemoveUnrelatedInstallOperations;
+use Php\Pie\ComposerIntegration\Listeners\RemoveUnrelatedInstallOperations;
 use Php\Pie\DependencyResolver\RequestedPackageAndVersion;
 use Php\Pie\Platform\Architecture;
 use Php\Pie\Platform\OperatingSystem;
